Build Terra, Physics, Sound, Network (mostly) under CMake/clang
Bring the cross-platform build frontier from Util/AI up through Terra,
Physics, and Sound, with Network nearly complete. The Game/Render header
chain now parses cross-platform.

Build infrastructure:
- Reconstruct the original precompiled-header set in the shared
  Util/XMath/StdAfx.h (xmath, Profiler, DebugPrm, DebugUtil, SystemUtil)
  for the header-less modules (AI, Terra, Physics, ...).
- Add include roots: Physics, Terra, Game, UserInterface, Network and the
  Render subdirs (inc/src/3dx/D3D) so the upper-layer header chain resolves.
- Serialization: simplify serializePolymorphic to a non-const T*& parameter
  (drop the MSVC-only const_cast); add a std::vector<bool> overload to avoid
  libc++'s proxy-reference breaking the generic element path.

Platform shims (Track B backends stubbed, not ported):
- Expand fake dsound.h / mmsystem.h / dplay8.h; add fake Lmcons.h.
- Add Win32 shim stubs (MK_* flags, WNDPROC, GetComputerName/GetUserName,
  MAX_COMPUTERNAME_LENGTH) to Platform/WindowsAPI.h.
- Gate DirectSound (SoundSystem/WaveFile) and DirectPlay
  (P2P_interface{1Th,3Th,2ThDPF,Aux}) backends behind if(WIN32); real SDL
  audio/networking come later.

Per-file clang fixes: const-ref temporaries (sTorpedo::quant,
FormationController::moveAction, StringInWrapper), enum fixed-underlying-type
(eGameOrder, eNetMessageCode, DirectControlMode), const comparison operators,
typename for dependent types, max() overload disambiguation, and three
corrupted "#include FileUtils/XGUID.h" lines missing the #include.

Exclude orphan sources not present in Core.vcproj/Game.vcproj (pre-refactor
leftovers referencing dead classes/headers): Physics/{Missile,
MultiBodyDispatcher,Rocket,PathTracking}, Network/{CommonEvents,GS_interface,
ConnectionDP}, and the standalone Sound/TextConverter tool. Physics
CrashSystem/RigidBodyNode remain deferred pending a stub Render library.

+++ Physics/RigidBodyMissile.cpp
@@ -51,7 +51,7 @@ void RigidBodyMissile::EulerEvolve(float dt)
 	quat.normalize();
 
 	// Linear damping anisotropic - apply in local frame
-	// Анизатропный дампинг приводит к повороту вектора скорости, при пастрекинге это вредно. Юнит разгоняется.
+	// пїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅ пїЅпїЅпїЅпїЅпїЅпїЅпїЅ пїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅ пїЅ пїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅ пїЅпїЅпїЅпїЅпїЅпїЅпїЅ пїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅ, пїЅпїЅпїЅ пїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅ пїЅпїЅпїЅ пїЅпїЅпїЅпїЅпїЅпїЅ. пїЅпїЅпїЅпїЅ пїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅ.
 	Vect3f damping(velocity_);
 	orientation().invXform(damping);
 	damping.x *= linear_damping.x*dt;
@@ -198,8 +198,8 @@ int RigidBodyMissile::checkCollisionPointed(const Vect3f& pos)
 int RigidBodyMissile::heightWithWater(int x, int y)
 {
 	if(waterAnalysis())
-		return max(vMap.gVBuf[vMap.offsetGBufC(x >> kmGrid, y >> kmGrid)], 
-			round(environment->water()->GetZ(x, y)));
+		return max(int(vMap.gVBuf[vMap.offsetGBufC(x >> kmGrid, y >> kmGrid)]),
+			(int)round(environment->water()->GetZ(x, y)));
 	else
 		return vMap.gVBuf[vMap.offsetGBufC(x >> kmGrid, y >> kmGrid)];
 }
